+++ src/Rtp/RtpProcess.cpp
@@ -139,7 +139,7 @@ bool RtpProcess::alive() {
         return true;
     }
 
-    GET_CONFIG(int, timeoutSec, RtpProxy::kTimeoutSec)
+    GET_CONFIG(uint64_t, timeoutSec, RtpProxy::kTimeoutSec)
     if (_last_frame_time.elapsedTime() / 1000 < timeoutSec) {
         return true;
     }
